+++ src/components/InformativeDialog.jsx
